LayoutTests/fast/mediastream/RTCPeerConnection-inspect-offer-expected.txt

Inspect the result of RTCPeerConnection.createOffer()

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Got offer
=== RTCSessionDescription ===
type: offer, sdp:
v=0
o=- {session-id:OK} 0 IN IP4 127.0.0.1
s=-
t=0 0
a=msid-semantic:WMS {media-stream-id:OK}
m=audio 9 UDP/TLS/RTP/SAVPF 111 8 0
c=IN IP4 0.0.0.0
a=rtcp-mux
a=sendrecv
a=rtpmap:111 OPUS/48000/2
a=rtpmap:8 PCMA/8000
a=rtpmap:0 PCMU/8000
a=ssrc:{ssrc:OK} cname:{cname:OK}
a=msid:{media-stream-id:OK} {media-stream-track-id:OK}
a=ice-ufrag:{ice-ufrag:OK}
a=ice-pwd:{ice-password:OK}
a=fingerprint:sha-256 8B:87:09:8A:5D:C2:F3:33:EF:C5:B1:F6:84:3A:3D:D6:A3:E2:9C:17:4C:E7:46:3B:1B:CE:84:98:DD:8E:AF:7B
a=setup:actpass
===

PASS Got offer
=== RTCSessionDescription ===
type: offer, sdp:
v=0
o=- {session-id:OK} 1 IN IP4 127.0.0.1
s=-
t=0 0
a=msid-semantic:WMS {media-stream-id:OK}
m=audio 9 UDP/TLS/RTP/SAVPF 111 8 0
c=IN IP4 0.0.0.0
a=rtcp-mux
a=sendrecv
a=rtpmap:111 OPUS/48000/2
a=rtpmap:8 PCMA/8000
a=rtpmap:0 PCMU/8000
a=ssrc:{ssrc:OK} cname:{cname:OK}
a=msid:{media-stream-id:OK} {media-stream-track-id:OK}
a=ice-ufrag:{ice-ufrag:OK}
a=ice-pwd:{ice-password:OK}
a=fingerprint:sha-256 8B:87:09:8A:5D:C2:F3:33:EF:C5:B1:F6:84:3A:3D:D6:A3:E2:9C:17:4C:E7:46:3B:1B:CE:84:98:DD:8E:AF:7B
a=setup:actpass
m=video 9 UDP/TLS/RTP/SAVPF 103 100 120
c=IN IP4 0.0.0.0
a=rtcp-mux
a=sendrecv
a=rtpmap:103 H264/90000
a=rtpmap:100 VP8/90000
a=rtpmap:120 RTX/90000
a=fmtp:103 packetization-mode=1
a=fmtp:120 apt=100;rtx-time=200
a=rtcp-fb:100 nack
a=rtcp-fb:103 nack pli
a=rtcp-fb:100 nack pli
a=rtcp-fb:103 ccm fir
a=rtcp-fb:100 ccm fir
a=ssrc:{ssrc:OK} cname:{cname:OK}
a=msid:{media-stream-id:OK} {media-stream-track-id:OK}
a=ice-ufrag:{ice-ufrag:OK}
a=ice-pwd:{ice-password:OK}
a=fingerprint:sha-256 8B:87:09:8A:5D:C2:F3:33:EF:C5:B1:F6:84:3A:3D:D6:A3:E2:9C:17:4C:E7:46:3B:1B:CE:84:98:DD:8E:AF:7B
a=setup:actpass
===

PASS End of promise chain
PASS successfullyParsed is true

TEST COMPLETE
